Restructure src/components/Tasks dir for cleaner code
