+++ src/components/Tasks/TaskBoard/styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    marginHorizontal: 20,
+    justifyContent: 'space-between',
+  },
+});
+
+export default styles;
